some work

+++ server/src/main/kotlin/net/starlegacy/feature/tutorial/message/PopupMessage.kt
@@ -17,7 +17,7 @@ open class PopupMessage(private val title: Component = text(""), private val sub
 				subtitle,
 				Title.Times.times(
 					Duration.of(10, Tick.tick()),
-					Duration.of(Long.MAX_VALUE - 20, Tick.tick()),
+					Duration.of(Int.MAX_VALUE - 20L, Tick.tick()),
 					Duration.of(0,  Tick.tick()))
 			)
 		)
